dapps/requirements.txt
Flask==2.0.3
requests
qrcode[pil]
speechrecognition
pycryptodome
numpy